Move CalculateAmountPay to Handler
Move CalculateAmountPay to Handler
Controller calls Handler instead of Database
Add GetCostKwh to Database
