+++ EnergiaClienteWebApi/Databases/EnergiaClienteDatabase.cs
@@ -1,4 +1,4 @@
-namespace EnergiaClienteWebApi.Databases.EnergiaClienteDatabase;
+namespace EnergiaClienteWebApi.Databases;
 
 using System.Data;
 using System.Text;
@@ -39,6 +39,8 @@ public record Invoice
     public byte[]? Document { get; set; }
 }
 
+public record CostKwh(decimal costkwhPonta,decimal costkwhCheias,decimal costkwhVazio);
+
 public class dbResponse<T>
 {
     public dbResponse()
@@ -55,22 +57,23 @@ public class StatusObject
     public StatusObject(bool _error)
     {
         this.Error = _error;
-        this.StatusCode=200;
+        this.StatusCode = 200;
     }
     public bool Error { get; set; }
     public int StatusCode { get; set; }
     public string? ErrorMessage { get; set; }
 }
 
-public class EnergiaCliente
+public class EnergiaClienteDatabase
 {
-    //private static SqlConnection connection = new SqlConnection("Data Source=MSIRUBEN\\SQLEXPRESS;Initial Catalog=EnergiaClienteDados;Integrated Security=True;TrustServerCertificate=True");
-
     private static SqlConnection connection = new SqlConnection("Data Source=192.168.1.8,1433;Initial Catalog=EnergiaClienteDados;User ID=sa;Password=1Secure*Password1;TrustServerCertificate=True");
+    
     //get from db
-    private static decimal costkwhPonta => 0.24m;
-    private static decimal costkwhCheias => 0.1741m;
-    private static decimal costkwhVazio => 0.1072m;
+    private static CostKwh costKwh => new CostKwh(0.24m,0.1741m,0.1072m);
+
+    public static CostKwh GetCostKwh(){
+        return costKwh;
+    }
 
     public static dbResponse<Invoice> GetInvoices(GetInvoicesRequestModel requestModel)
     {
@@ -144,19 +147,6 @@ public class EnergiaCliente
             Result = readings,
             Status = new StatusObject(false)
         };
-    }
-
-    public static dbResponse<decimal> CalculateAmountPay(int habitation, Reading reading)
-    {
-        var readings = GetReadings(new GetReadingsRequestModel() { habitation = habitation, quantity = 1 });
-
-        if (readings.Result == null) return new dbResponse<decimal>() { Status = readings.Status };
-
-        var last = readings.Result[0];
-
-        decimal amount = (reading.Ponta - last.Ponta) * costkwhPonta + (reading.Cheias - last.Cheias) * costkwhCheias + (reading.Vazio - last.Vazio) * costkwhVazio;
-
-        return new dbResponse<decimal>() { Result = new List<decimal> { amount } };
     }
 
     private static DataRowCollection RunSelectProcedure(string procedure, SqlParameter[] parameters)
